+++ elevation.js
@@ -8,6 +8,7 @@ getProperty,
 Hooks,
 PointSource,
 Token,
+VisionSource,
 Wall
 */
 /* eslint no-unused-vars: ["error", { "argsIgnorePattern": "^_" }] */
@@ -43,8 +44,16 @@ lights can display with varying canvas elevation.
 
 */
 
+function addHook(name, fn) {
+  const hooks = CONFIG.GeometryLib.hooks ??= new Map();
+  const id = Hooks.on(name, fn);
+  hooks.set(name, id);
+  return id;
+}
+
 export function registerElevationAdditions() {
-  if ( CONFIG.GeometryLib.proneStatusId ) return; // Already registered.
+  CONFIG.GeometryLib.registered ??= new Set();
+  if ( CONFIG.GeometryLib.registered.has("Elevation") ) return;
 
   // Define elevation getters.
   // Because elevation is saved to flags, use an async method instead of a setter.
@@ -78,9 +87,9 @@ export function registerElevationAdditions() {
 
   // Tile
   // Sync tile.document.elevation with tile.document.flags.elevatedvision.elevation
-  Hooks.on("preUpdateTile", preUpdateTileHook);
-  Hooks.on("updateTile", updateTileHook);
-  Hooks.on("drawTile", drawTileHook);
+  addHook("preUpdateTile", preUpdateTileHook);
+  addHook("updateTile", updateTileHook);
+  addHook("drawTile", drawTileHook);
 
   // Token
   addClassGetter(Token.prototype, "elevationE", tokenElevationE);
@@ -105,7 +114,7 @@ export function registerElevationAdditions() {
 
   // Sync token.tokenHeight between EV and Wall Height
   // Also clear the _tokenHeight cached property.
-  Hooks.on("preUpdateToken", preUpdateTokenHook);
+  addHook("preUpdateToken", preUpdateTokenHook);
 
   // Wall
   addClassGetter(Wall.prototype, "topE", wallTopE);
@@ -119,7 +128,9 @@ export function registerElevationAdditions() {
   addClassMethod(Wall.prototype, "setBottomZ", setZBottom);
 
   // Sync wall bottom and top elevations between EV and Wall Height
-  Hooks.on("preUpdateWall", preUpdateWallHook);
+  addHook("preUpdateWall", preUpdateWallHook);
+
+  CONFIG.GeometryLib.registered.add("Elevation");
 }
 
 /* Elevation handling
@@ -182,7 +193,7 @@ function visionSourceElevationE() {
   return this.object?.topE ?? this.object?.elevationE ?? this.data.elevation ?? 0;
 }
 
-async function setVisionSourceElevationE(value) {
+async function setVisionSourceElevationE(_value) {
   console.warn("Cannot set elevationE for a vision source because it is calculated from token height.");
   return;
 }
@@ -295,8 +306,8 @@ function preUpdateTileHook(_tileD, changes, _options, _userId) {
   const evFlag = MODULE_KEYS.EV.FLAG_PLACEABLE_ELEVATION;
   const updates = {};
   if ( changeKeys.has(evFlag) ) updates.elevation = flatData[evFlag];
-  else if ( changeKeys.has("elevation") ) updates[evFlag] = data.elevation;
-  foundry.utils.mergeObject(data, updates);
+  else if ( changeKeys.has("elevation") ) updates[evFlag] = changes.elevation;
+  foundry.utils.mergeObject(changes, updates);
 }
 
 function updateTileHook(tileD, changed, _options, _userId) {